LogSplitter use PathBuf instead of String
Implement ensure_correct_file_extension function that removes all file
extensions and then adds `.xes` or `.xes.gz` appropriately.
In general it would be better to just remove the suffix `.gz` and then
the suffix `.xes` instead of removing _all_ extensions. This is not that
important at the moment though